+++ src/Main/java/Tickets/ConcertTicket.java
@@ -7,7 +7,7 @@ import java.util.Date;
 import java.util.HashMap;
 
 public class ConcertTicket extends Ticket{
-    public ConcertTicket(String description, HashMap<User, Integer> users, int value, String purchaseDate, int splitEvenly){
+    public ConcertTicket(String description, HashMap<User, Double> users, double value, String purchaseDate, int splitEvenly){
         super(description, users, value, purchaseDate, splitEvenly);
     }
 }
