+++ src/Main/java/Tickets/TaxiTicket.java
@@ -7,7 +7,7 @@ import java.util.Date;
 import java.util.HashMap;
 
 public class TaxiTicket extends Ticket{
-    public TaxiTicket(String description, HashMap<User, Integer> users, int value, String purchaseDate, int splitEvenly){
+    public TaxiTicket(String description, HashMap<User, Double> users, double value, String purchaseDate, int splitEvenly){
         super(description, users, value, purchaseDate, splitEvenly);
     }
 }
